+++ scripts/all-combined.js
@@ -19,6 +19,7 @@ module.exports = {
 		'../rules/best-practices.js',
 		'../rules/ecmascript-6.js',
 		'../rules/ecmascript-8.js',
+		'../rules/ecmascript-9.js',
 		'../rules/node-and-common.js',
 		'../rules/plugin-babel.js',
 		'../rules/plugin-chai-expect.js',